backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
matplotlib>=3.8.0
google-generativeai>=0.3.0
python-multipart>=0.0.6 
python-dotenv